+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.3.1 (Next)
 - Fix GPG keys output path message.
 - Fix encrypted data validation.
+- Enforce non-empty password when input.
 
 ### 1.3.0
 - Bump minimum Ruby version to 2.6.